requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
PyPDF2==3.0.1
pikepdf==8.10.1
python-docx==1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
